Warning消し
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@5676 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDConvexMeshInterpolate.cpp
@@ -34,7 +34,7 @@ Vec3f CDConvexMeshInterpolate::Support(const Vec3f& v) const {
 	while (1) {
 		const std::vector<int>& curNeighbor = neighbor[curPos];
 		int i = 0;
-		int n = curNeighbor.size();
+		int n = (int)curNeighbor.size();
 		while(i!=n){
 			++ count;
 			if (curNeighbor[i] == lastPos){
